Revert to using prettyPhoto 3.0b.
The 3.0 release version of prettyPhoto was
showing unreliable behavior in closing a
Vimeo video popup, especially with clicking the
'X' close icon.  The 3.0b version does not
have this problem.

If the bug can be narrowed down to either a code
bug in prettyPhoto or with using the Vimeo
Universal player, then a workaround can be used
to move to the release version of prettyPhoto.