Refactor portfolio detail pages to have fewer places to modify per page.
Previously, in order to make the page navigation
appear correctly for each portfolio detail page,
it was necessary to remember to set ".active" and
".selected" CSS classes on the sidebar nav
elements for that page.  This could lead to more
error-prone copy/paste when creating new detail
pages.

Now, the nav HTML is identical, and the display is
dictated by body CSS classes and JS.